pyproject.toml
[project]
name = "apex-env"
version = "1.0.0"
description = "APEX Professional Tasks Environment for RL training"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.115.0",
    "groq>=0.11.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.32.0",
    "taskipy>=1.14.0",
    "trl>=0.29.0",
    "transformers>=5.3.0",
    "datasets>=4.6.1",
    "accelerate>=1.13.0",
    "peft>=0.18.1",
    "torch>=2.10.0",
]

[tool.taskipy.tasks]
serve = "python main.py"
test = "python -m pytest test/ -v"
